Move authentication to it's own View Controller.
I was admittedly trying some fanciness by putting the LoginView and
RegistrationView in the CameraViewController and showing them both
with a translucent background (so you'd see the camera output behind it).
This proved to be visually harmful and made the CameraViewController
carry more burden than it should have.  So, now it's gone and auth
has it's own view controller.

A side effect of this change is that the AuthenticationViewController
needs to update the HAL Resource carried by the CameraViewController
before it is dismissed.  I'm not wild about how this is currently done,
but it works.