Localization.
I'm including the context in the localization key. For example, the 
English string "Username" in the registration view has the key 
"authentication: username".  It seems that the typical way of doing
this is to just use the english word as the string, but this seems
inelegant and prone to oddness when text changes at some point in the
future.  In other words, I'd rather have a semantic key than pretend
that the key will match the English translation.