Register device for push notifications.
We're using an observer in the CameraViewController and AppDelegate to
handle hitting the API when the registration callback is fired off (in
the AppDelegate).  I'm not wild about this as notifications have been a
pain point in the past, but we need the root resource from the
CameraViewController in order to get the HAL link for the device
registration.